requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
PyJWT==2.8.0
cryptography==41.0.7
tensorflow==2.15.0
Pillow==10.1.0
numpy==1.24.3 